feat: make a builder pattern for txt2img
